Add `systemd-extras` to `polkit` - provides several polkit rules.
Various elevated privileges are available to logged in users (those
on the console with physical access) and allow that users to perform
restarts and other basic system functions.

Adding these here allows this to work for console users on systems
that do not have a desktop installed. For all current installs with
polit, systemd-extras is already part of the bundle.

Since polkit is not a default bundle, this means no change for most
server installs and no default change.

Addresses #1633