desktop: replace also-add for include bundle desktop-assets-extras
This fixes https://github.com/clearlinux/distribution/issues/1766

While we want to decouple 3rd party theme/extensions, we don't
want to break defaults like this, marking desktop-assets-extras
as a soft requirement won't install it when updating, but on new
bundle-add commands only.

This change should be reverted at some point in the future (probaby
in the middle of a format bump).

Signed-off-by: Simental Magana, Marcos <marcos.simental.magana@intel.com>