Remove R-treatSens from R-extras bundle
The latest release of the `R-treatSens` package was made on 2018-03-09,
and the package was removed from CRAN more than one year ago with this
notice:

    Archived on 2019-04-07 as serious check problems were not corrected
    despite a reminder.

The package has also been failing to build for some time. If it is later
added back to CRAN and receiving active maintenance, we can consider
adding it back.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>