Remove R-R2Cuba from R-extras bundle
The latest release of the `R-R2Cuba` package was made on 2015-10-02, and
the package was removed from CRAN more than one year ago with this
notice:

    Archived on 2018-07-15 after over a month's notice to revdeps as has
    serious errors.

When GCC 10 was merged, the package build began to fail, so I think the
best course of action for now is to remove it. No other CRAN packages in
the repo depend on it.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>